+++ test/Hpack/UtilSpec.hs
@@ -3,12 +3,10 @@
 module Hpack.UtilSpec (main, spec) where
 
 import           Data.Aeson
-import           Data.Aeson.QQ
 import           Data.Aeson.Types
 import           Helper
 import           System.Directory
 
-import           Hpack.Config
 import           Hpack.Util
 
 main :: IO ()
@@ -72,26 +70,23 @@ spec = do
           getModuleFilesRecursive "foo" `shouldReturn` empty
 
   describe "List" $ do
-    let invalid = [aesonQQ|{
-          name: "hpack",
-          gi: "sol/hpack",
-          ref: "master"
-        }|]
-        parseError :: String -> Either String (List Dependency)
-        parseError prefix = Left (prefix ++ ": neither key \"git\" nor key \"github\" present")
+    let
+      parseError :: String -> Either String (List Int)
+      parseError prefix = Left (prefix ++ ": expected Int, encountered String")
+
     context "when parsing single values" $ do
       it "returns the value in a singleton list" $ do
         fromJSON (toJSON $ Number 23) `shouldBe` Success (List [23 :: Int])
 
       it "returns error messages from element parsing" $ do
-        parseEither parseJSON invalid `shouldBe` parseError "Error in $"
+        parseEither parseJSON (String "foo") `shouldBe` parseError "Error in $"
 
     context "when parsing a list of values" $ do
       it "returns the list" $ do
         fromJSON (toJSON [Number 23, Number 42]) `shouldBe` Success (List [23, 42 :: Int])
 
       it "propagates parse error messages of invalid elements" $ do
-        parseEither parseJSON (toJSON [String "foo", invalid]) `shouldBe` parseError "Error in $[1]"
+        parseEither parseJSON (toJSON [Number 23, String "foo"]) `shouldBe` parseError "Error in $[1]"
 
   describe "tryReadFile" $ do
     it "reads file" $ do